standardize PositionName / FOV name
