+++ src/main/scala/package.scala
@@ -11,15 +11,18 @@ import scopt.Read
 
 import upickle.default.{ Reader as JsonReader }
 import io.github.iltotore.iron.{ :|, refineEither }
-import io.github.iltotore.iron.Not
+import io.github.iltotore.iron.constraint.any.{ StrictEqual, Not }
+import io.github.iltotore.iron.constraint.string.Match
 import io.github.iltotore.iron.constraint.collection.{ Empty, ForAll }
 import io.github.iltotore.iron.constraint.char.*
 
 import at.ac.oeaw.imba.gerlich.gerlib.SimpleShow
 import at.ac.oeaw.imba.gerlich.gerlib.imaging.{ ImagingTimepoint, PositionName }
+import at.ac.oeaw.imba.gerlich.gerlib.imaging.instances.all.given
 import at.ac.oeaw.imba.gerlich.gerlib.json.JsonValueWriter
 import at.ac.oeaw.imba.gerlich.gerlib.numeric.*
 import at.ac.oeaw.imba.gerlich.gerlib.numeric.instances.nonnegativeInt.given // Order, Show
+import at.ac.oeaw.imba.gerlich.gerlib.syntax.all.*
 import at.ac.oewa.imba.gerlich.looptrace.RowIndexAdmission
 import at.ac.oeaw.imba.gerlich.looptrace.syntax.json.*
 
@@ -185,4 +188,20 @@ package object looptrace {
         def fromRoiIndex(i: RoiIndex): TraceId = new TraceId(i.get)
         def unsafe = NonnegativeInt.unsafe.andThen(TraceId.apply)
     end TraceId
+
+    // A 1-based count encoding of the field of view, prefixed with "P" for "position", and 4 digits to hold up to 9999 FOVs
+    private type OneBasedFourDigitPositionNameConstraint = Match["P\\d{4}"] & Not[StrictEqual["P0000"]]
+    
+    // A String which complies with the domain restriction for representing a field of view by a 1-based count
+    type OneBasedFourDigitPositionName = String :| OneBasedFourDigitPositionNameConstraint
+    
+    /** Helpers for working with the String refinement representing field of view name */
+    object OneBasedFourDigitPositionName:
+        /** Attempt to further refine the position name as one in compliance with using 4 digits for a one-based count */
+        def fromPositionName(trimZarr: Boolean): PositionName => Either[String, OneBasedFourDigitPositionName] = 
+            ((_: PositionName).show_) `andThen` fromString(trimZarr)
+        
+        def fromString(trimZarr: Boolean): String => Either[String, OneBasedFourDigitPositionName] = 
+            s => (if trimZarr then s.stripSuffix(".zarr") else s).refineEither[OneBasedFourDigitPositionNameConstraint]
+    end OneBasedFourDigitPositionName
 }
